+++ ide/CHANGELOG.md
@@ -11,6 +11,9 @@
   useful statistics.
 - [Signed builds.][1366] Our builds are signed and will avoid warnings from the
   operating system about being untrusted.
+- [Area selection][1588]. You can now select multiple nodes at once. Just click
+  and drag on the background of your graph and see the beauty of the area
+  selection appear.
 - [Opening projects in application graphical interface][1587]. Press `cmd`+`o`
   to bring the list of projects. Select a project on the list to open it.
 - [Initial support for undo-redo][1602]. Press <kbd>cmd</kbd>+<kbd>z</kbd> to
@@ -29,6 +32,7 @@
 
 #### Enso Compiler
 
+[1588]: https://github.com/enso-org/ide/pull/1588
 [1577]: https://github.com/enso-org/ide/pull/1577
 [1587]: https://github.com/enso-org/ide/pull/1587
 [1366]: https://github.com/enso-org/ide/pull/1366